pyproject.toml
[tool.poetry]
name = "course-job-5"
version = "0.1.0"
description = ""
authors = ["gelena <helenatcitok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"
black = "^24.1.0"
requests = "^2.31.0"
psycopg = "^3.1.17"
binary = "^1.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
